refactor: store migration
