+++ app/page.tsx
@@ -1,10 +1,5 @@
-/* eslint-disable react-refresh/only-export-components */
 import { redirect } from 'next/navigation';
 import './globals.css';
-
-// export function generateStaticParams() {
-//   return [{ slug: [''] }];
-// }
 
 export default function Page() {
   redirect('/character');
