+++ src/app/index.ts
@@ -1,5 +1,5 @@
 export {
-  mainStore,
+  makeStore,
   type AppStore,
   type RootState,
   type AppDispatch,
