+++ src/app/providers/StoreProvider/StoreProvider.tsx
@@ -1,5 +1,7 @@
-import type { ReactNode } from 'react';
-import { mainStore } from '../../stores';
+'use client';
+
+import { useRef, type ReactNode } from 'react';
+import { type AppStore, makeStore } from '../../stores';
 import { Provider } from 'react-redux';
 
 type Props = {
@@ -7,5 +9,10 @@ type Props = {
 };
 
 export const StoreProvider = ({ children }: Props) => {
-  return <Provider store={mainStore}>{children}</Provider>;
+  const storeRef = useRef<AppStore>(undefined);
+  if (!storeRef.current) {
+    storeRef.current = makeStore();
+  }
+
+  return <Provider store={storeRef.current}>{children}</Provider>;
 };
