+++ src/app/stores/store.ts
@@ -2,15 +2,17 @@ import { configureStore } from '@reduxjs/toolkit';
 import { selectSlice } from '@features';
 import { startrackApi } from '../../entities/character/api/startrack';
 
-export const mainStore = configureStore({
-  reducer: {
-    select: selectSlice.reducer,
-    [startrackApi.reducerPath]: startrackApi.reducer,
-  },
-  middleware: (getDefaultMiddleware) =>
-    getDefaultMiddleware().concat(startrackApi.middleware),
-});
+export const makeStore = () => {
+  return configureStore({
+    reducer: {
+      select: selectSlice.reducer,
+      [startrackApi.reducerPath]: startrackApi.reducer,
+    },
+    middleware: (getDefaultMiddleware) =>
+      getDefaultMiddleware().concat(startrackApi.middleware),
+  });
+};
 
-export type AppStore = typeof mainStore;
+export type AppStore = ReturnType<typeof makeStore>;
 export type RootState = ReturnType<AppStore['getState']>;
 export type AppDispatch = AppStore['dispatch'];
