+++ src/components/ThemeToggle/ThemeToggle.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useTheme } from '../../app';
 
 export const ThemeToggle = () => {
