+++ src/entities/character/api/index.ts
@@ -1,3 +1,5 @@
+export { getInitialCharacters } from './get-initial-characters';
+
 export {
   startrackApi,
   useGetCharacterQuery,
