+++ src/shared/index.ts
@@ -1,5 +1,5 @@
 export { SEARCH_KEY, CHAR_MAP } from './constants';
 export { cn } from './lib';
-export { Button, MsgBlock } from './ui-kit';
+export { Button, MsgBlock, Navigation } from './ui-kit';
 export { ResponsePage } from './api';
 export { useLocalStorage } from './hooks';
