+++ src/shared/ui-kit/Navigation/index.ts
@@ -1,0 +1,1 @@
+export { Navigation } from './Navigation';
