+++ src/shared/ui-kit/index.ts
@@ -1,2 +1,3 @@
 export { Button } from './Button';
 export { MsgBlock } from './MsgBlock';
+export { Navigation } from './Navigation';
